+++ packages/core/src/components/screens/release-notes/WhatsNewScreen.tsx
@@ -2,16 +2,18 @@ import React, { useEffect, useState } from "react";
 import { ChangeLog, Release } from "./ChangeLog";
 import * as mockData from "../../../static/mock-data/whats-new.json";
 import { useAppVersion, useEnvironment, useServicePortalName } from "../../../store/mad-config";
-import { useReleaseNotesVersion } from "../../../store/release-notes/release-notes";
-import { CircularProgress } from "@equinor/mad-components";
+import { useReleaseNotesVersion } from "../../../store/release-notes";
+import {Button, CircularProgress, EDSStyleSheet, useStyles} from "@equinor/mad-components";
 import { useCoreStackNavigation } from "../../../hooks/useCoreStackNavigation";
 import { useDemoMode } from "../../../store/demo-mode";
 import { fetchReleaseNotes } from "./fetchReleaseNotes";
+import { View } from "react-native";
 
 /**
  * This screen will display the latest releasenotes
  */
 export const WhatsNewScreen = () => {
+    const styles = useStyles(whatsNewStyles);
     const environment = useEnvironment();
     const releaseNotesVersion = useReleaseNotesVersion();
     const servicePortalName = useServicePortalName();
@@ -43,16 +45,46 @@ export const WhatsNewScreen = () => {
     }
 
     if (!release) {
-        return <CircularProgress />;
+        return (
+            <View style={styles.spinnerContainer}>
+                <CircularProgress />
+            </View>
+        );
     }
 
     return (
-        <ChangeLog
-            release={release}
-            onPressAffirm={() => {
-                releaseNotesVersion.setLastDisplayedReleaseNotesVersion(appVersion);
-                navigate();
-            }}
-        />
+        <View style={styles.container}>
+            <ChangeLog release={release} />
+            <View style={styles.footer}>
+                <Button
+                    title="OK"
+                    onPress={() => {
+                        releaseNotesVersion.setLastDisplayedReleaseNotesVersion(appVersion);
+                        navigate();
+                    }}
+                    style={{ width: 81 }}
+                />
+            </View>
+        </View>
     );
 };
+
+const whatsNewStyles = EDSStyleSheet.create((theme) => ({
+    spinnerContainer: {
+        display: "flex",
+        height: "100%",
+        justifyContent: "center",
+        alignItems: "center"
+    },
+    container: {
+        display: "flex",
+        paddingTop: theme.geometry.dimension.cell.minHeight,
+        height: "100%",
+        justifyContent: "space-between",
+    },
+    footer: {
+        alignItems: "flex-end",
+        marginHorizontal: theme.spacing.container.paddingHorizontal,
+        marginVertical: theme.spacing.container.paddingVertical,
+    },
+}));