+++ packages/navigation/src/_internal/getCustomRenderFunction.tsx
@@ -1,5 +1,6 @@
 import React, { ReactNode } from "react";
 import { HeightSender } from "../header-height-context";
+import { View } from "react-native";
 /**
  * Custom render function that displays an environment banner
  * @param originalRender the original render function, found in the descriptor
@@ -14,7 +15,7 @@ export const getCustomRenderFunction = (
         <>
             {CustomSubHeader && <CustomSubHeader />}
             <HeightSender />
-            {originalRender()}
+            <View style={{ flex: 1 }}>{originalRender()}</View>
         </>
     );
     return customRender;